ISSUE #3787 - move two toned icons
