+++ frontend/assets/icons/twoToned/remove_element-two_toned.svg.tsx
