+++ frontend/assets/icons/twoToned/warning_small-two_toned.svg.tsx
